D Desktop | sidebar: not connectable when no name other: moved unused stuff to .dep folder
