+++ desktop/src/main/kotlin/com/github/picture2pc/desktop/ui/main/elements/dep/Throbbers.kt
@@ -1,4 +1,4 @@
-package com.github.picture2pc.desktop.ui.main.elements
+package com.github.picture2pc.desktop.ui.main.elements.dep
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
